mis les interfaces dans des dossiers externes et supprimé du code mort.
