+++ client/src/types/AddRecipe.d.ts
@@ -1,0 +1,24 @@
+export interface RecipeData {
+  titre: string;
+  recette_ref: string;
+  image_url: string;
+  description: string;
+  temps_id: number;
+  difficulte_id: number;
+  type_id: number;
+  preparation: { id: string; ordre: number; description: string }[];
+  saison: string;
+  utilisateur_id: number;
+}
+
+export interface Ingredient {
+  nom: string;
+  id: string;
+  icone_categorie: string;
+  unite: string;
+}
+
+export interface IngredientData extends Ingredient {
+  quantite: number;
+  recette_ref: string;
+}
